got calculateMarkup returning an int by using casting to reflect a percentage value

+++ src/main/java/behaviours/ISell.java
@@ -1,5 +1,5 @@
 package behaviours;
 
 public interface ISell {
-    public double calculateMarkUp();
+    public int calculateMarkUp();
 }
